requirements.txt
langchain
chromadb==0.3.22
llama-cpp-python
urllib3
InstructorEmbedding
sentence-transformers==2.2.2
faiss-cpu
huggingface_hub
transformers
protobuf
accelerate
bitsandbytes
python-dotenv~=1.0.0
tqdm~=4.65.0
torch~=2.0.1
einops
xformers
peft
pyperclip